updating smooth fps
