+++ config.ini
@@ -18,7 +18,7 @@ bitrate = 0 ; Sets the bitrate of the video in KB/s. If bitrate is set as 0, the
 ; s3 prefixs for input/output:
 method_name = fps_bitrate
 input_prefix_s3 = original-videos/
-output_prefix_s3 = reduced-videos/${method_name}-fps_${fps}-bitrate_${bitrate}/
+output_prefix_s3 = reduced-videos/${method_name}-fps_${fps_factor}x-bitrate_${bitrate}/
 temp_path = reduced-videos/ ; A temp folder to store video for uploading to s3
 
 
@@ -55,16 +55,16 @@ codec = mp4v ; desired video codec
 
 ; cloud_functionality parameters:
 download_model = True ; boolean to indicate if a model needs to be downloaded
-local_model_path = fsrcnn_x4.pb ; local path to save pre-trained model
+local_model_path = smooth.pkl ; local path to save pre-trained model
 clean_model = True ; boolean to indicate to clean video or not
 
 ; s3 prefixs for input/output paths and models:
-method_name = superres
-input_prefix_s3 = reduced-videos/${reduction.ffmpeg_resolution_downsampler:method_name}-quality_${reduction.ffmpeg_resolution_downsampler:quality}-algorithm_${reduction.ffmpeg_resolution_downsampler:algorithm}/
+method_name = smooth_fps
+
+
+input_prefix_s3 = reduced-videos/${reduction.fps_bitrate:method_name}-fps_${reduction.fps_bitrate:fps_factor}x-bitrate_${reduction.fps_bitrate:bitrate}/
 output_prefix_s3 = reconstructed-videos/${method_name}-codec_${codec}/
-model_prefix_s3 = pretrained-models/fsrcnn_x4.pb ; s3 prefix of the pre-trained model
-
-
+model_prefix_s3 = pretrained-models/smooth.pkl ; s3 prefix of the pre-trained model
 
 
 [reconstruction.fastsrgan]
@@ -124,8 +124,10 @@ clean_model = True ; boolean to indicate to clean video or not
 
 ; s3 prefixs for input/output paths and models:
 method_name = superres
-input_prefix_s3 = reduced-videos/${reduction.ffmpeg_resolution_downsampler:method_name}-quality_${reduction.ffmpeg_resolution_downsampler:quality}-algorithm_${reduction.ffmpeg_resolution_downsampler:algorithm}/
-output_prefix_s3 = reconstructed-videos/${method_name}-codec_${codec}/
+; input_prefix_s3 = reduced-videos/${reduction.ffmpeg_resolution_downsampler:method_name}-quality_${reduction.ffmpeg_resolution_downsampler:quality}-algorithm_${reduction.ffmpeg_resolution_downsampler:algorithm}/
+input_prefix_s3 = reduced-videos/fps_bitrate-fps_2x-bitrate_0/
+output_prefix_s3 = reconstructed-videos/${method_name}-codec_${codec}/fps-reduced/
+; output_prefix_s3 = reconstructed-videos/${method_name}-codec_${codec}/
 model_prefix_s3 = pretrained-models/fsrcnn_x4.pb ; s3 prefix of the pre-trained model
