+++ reconstruction/smooth_fps/smooth_fps.py
@@ -18,6 +18,7 @@ root_path = subprocess.run(
 sys.path.append(root_path)
 
 from utilities import CloudFunctionality
+from utilities import ConfigHandler
 
 
 
@@ -130,13 +131,12 @@ if __name__ == '__main__':
     cloud_functionality = CloudFunctionality()
 
     # load and allocate config file
-    config = configparser.ConfigParser(inline_comment_prefixes=';')
-    config.read('../../config.ini')
-    s3_args = config['DEFAULT']
-    method_args = config['reconstruction.recon_args']
-    logging.info("successfully loaded config file")
+    config = ConfigHandler('reconstruction.smooth_fps')
+    s3_args = config.s3
+    method_args = config.method
+    
 
-    cloud_functionality.preprocess(method_args, s3_args)
+    cloud_functionality.preprocess(s3_args, method_args)
 
     ####################### Check for errors in the args ###################################
